requirements.txt
djangorestframework
django==3.0.3
gunicorn==20.0.4
